Looks like the test code has moved!?
