+++ src/test/resources/trusted/foo.groovy
@@ -1,4 +1,4 @@
-import org.jenkinsci.plugins.workflow.CpsFlowExecutionTest
+import org.jenkinsci.plugins.workflow.cps.CpsFlowExecutionTest
 
 // try to touch a sensitive method, which may or may not fail
 // depending on the context we run in
